requirements.txt
openai==1.63.2
python-dotenv
PyYAML